@@ -1,89 +1,88 @@
 plugins {
-	id 'eclipse'
-	id 'maven-publish'
-	id 'net.minecraftforge.gradle' version "${forge_gradle_version}"
-	id 'org.parchmentmc.librarian.forgegradle' version "${parchent_librarian_version}.+"
-    id 'com.matthewprenger.cursegradle' version "${cursegradle_version}"
+    id 'java-library'
+    id 'idea'
+    id 'maven-publish'
+    id 'net.neoforged.gradle.userdev' version '7.0.165'
 }
-//apply plugin: 'net.minecraftforge.gradle'
-//apply plugin: 'org.parchmentmc.librarian.forgegradle'
-// Only edit below this line, the above code adds and enables the necessary things for Forge to be setup.
-//apply plugin: 'eclipse'
-//apply plugin: 'maven-publish'
-//apply from: 'https://raw.githubusercontent.com/SizableShrimp/Forge-Class-Remapper/main/classremapper.gradle'
+
+tasks.named('wrapper', Wrapper).configure {
+    // Define wrapper values here so as to not have to always do so when updating gradlew.properties.
+    // Switching this to Wrapper.DistributionType.ALL will download the full gradle sources that comes with
+    // documentation attached on cursor hover of gradle classes and methods. However, this comes with increased
+    // file size for Gradle. If you do switch this to ALL, run the Gradle wrapper task twice afterwards.
+    // (Verify by checking gradle/wrapper/gradle-wrapper.properties to see if distributionUrl now points to `-all`)
+    distributionType = Wrapper.DistributionType.BIN
+}
+
+group = mod_group_id
+version = "${minecraft_version}-" + mod_version
 
 libsDirName = "../output"
 distsDirName = "../output"
 
-group = blastcraft_group
-version = "${minecraft_version}-" + blastcraft_version
-
-base {
-
-	archivesName = blastcraft_name
-
+repositories {
+    mavenLocal()
 }
 
-java.toolchain.languageVersion = JavaLanguageVersion.of(17)
+base {
+    archivesName = mod_id
+}
+
+java.toolchain.languageVersion = JavaLanguageVersion.of(21)
 
 println('Java: ' + System.getProperty('java.version') + ' JVM: ' + System.getProperty('java.vm.version') + '(' + System.getProperty('java.vendor') + ') Arch: ' + System.getProperty('os.arch'))
-minecraft {
-    mappings channel: 'parchment', version: "${parchment_version}-${minecraft_version}"
-    runs {
-        client {
-            workingDirectory project.file('run')
 
-            property 'forge.logging.markers', 'SCAN,REGISTRIES,REGISTRYDUMP'
+runs {
 
-            property 'forge.logging.console.level', 'debug'
+    configureEach {
 
-            mods {
-                blastcraft {
-                    source sourceSets.main
-                }
-            }
-        }
+        systemProperty 'forge.logging.markers', 'REGISTRIES'
 
-        server {
-            workingDirectory project.file('run')
+        systemProperty 'forge.logging.console.level', 'debug'
 
-            // Recommended logging data for a userdev environment
-            property 'forge.logging.markers', 'SCAN,REGISTRIES,REGISTRYDUMP'
+        modSource project.sourceSets.main
+    }
 
-            // Recommended logging level for the console
-            property 'forge.logging.console.level', 'debug'
+    client {
 
-            mods {
-                blastcraft {
-                    source sourceSets.main
-                }
-            }
-        }
+        systemProperty 'forge.enabledGameTestNamespaces', project.mod_id
 
-        data {
-            workingDirectory project.file('run')
+    }
 
-            // Recommended logging data for a userdev environment
-            property 'forge.logging.markers', 'SCAN,REGISTRIES,REGISTRYDUMP'
+    server {
 
-            // Recommended logging level for the console
-            property 'forge.logging.console.level', 'debug'
+        systemProperty 'forge.enabledGameTestNamespaces', project.mod_id
+        programArgument '--nogui'
 
-            args '--mod', 'blastcraft', '--all', '--output', file('src/generated/resources/'), '--existing', file('src/main/resources/')
+    }
 
-            mods {
-                blastcraft {
-                    source sourceSets.main
-                }
-            }
-        }
+    gameTestServer {
+
+        systemProperty 'forge.enabledGameTestNamespaces', project.mod_id
+
+    }
+
+    data {
+
+        programArguments.addAll(
+                '--mod', project.mod_id,
+                '--all',
+                '--output', file('src/generated/resources/').getAbsolutePath(),
+                '--existing', file('src/main/resources/').getAbsolutePath(),
+                '--existing-mod', 'electrodynamics'
+        )
+
     }
 }
 
 sourceSets.main.resources { srcDir 'src/generated/resources' }
 
+configurations {
+    runtimeClasspath.extendsFrom localRuntime
+}
+
 dependencies {
-	def Folder = new File("C:/AmpzLibz")
+    def Folder = new File("C:/AmpzLibz")
     if(Folder.exists()) {
         implementation fileTree(Folder)
     }
