@@ -1,32 +1,29 @@
 package blastcraft;
 
 import blastcraft.client.ClientRegister;
-import blastcraft.common.recipe.BlastCraftRecipeInit;
+import blastcraft.common.block.BlastcraftBlockStates;
 import blastcraft.common.settings.Constants;
 import blastcraft.common.tag.BlastcraftTags;
 import blastcraft.registers.UnifiedBlastcraftRegister;
 import electrodynamics.prefab.configuration.ConfigurationHandler;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
-import net.minecraftforge.eventbus.api.IEventBus;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
-import net.minecraftforge.fml.common.Mod.EventBusSubscriber.Bus;
-import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
-import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
-import net.minecraftforge.fml.event.lifecycle.FMLLoadCompleteEvent;
-import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.api.distmarker.OnlyIn;
+import net.neoforged.bus.api.IEventBus;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.fml.common.EventBusSubscriber;
+import net.neoforged.fml.common.Mod;
+import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
+import net.neoforged.fml.event.lifecycle.FMLCommonSetupEvent;
+import net.neoforged.fml.event.lifecycle.FMLLoadCompleteEvent;
 
 @Mod(References.ID)
-@EventBusSubscriber(modid = References.ID, bus = Bus.MOD)
+@EventBusSubscriber(modid = References.ID, bus = EventBusSubscriber.Bus.MOD)
 public class Blastcraft {
 
-	public Blastcraft() {
+	public Blastcraft(IEventBus bus) {
 		ConfigurationHandler.registerConfig(Constants.class);
-		IEventBus bus = FMLJavaModLoadingContext.get().getModEventBus();
+		BlastcraftBlockStates.init();
 		UnifiedBlastcraftRegister.register(bus);
-		BlastCraftRecipeInit.RECIPE_SERIALIZER.register(bus);
 	}
 
 	@SubscribeEvent
