@@ -3,17 +3,20 @@ package blastcraft.client;
 import blastcraft.References;
 import blastcraft.client.guidebook.ModuleBlastcraft;
 import blastcraft.client.render.tile.RenderCamoflage;
-import blastcraft.registers.BlastcraftBlockTypes;
+import blastcraft.registers.BlastcraftFluids;
+import blastcraft.registers.BlastcraftTiles;
 import electrodynamics.client.guidebook.ScreenGuidebook;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
-import net.minecraftforge.client.event.EntityRenderersEvent;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
-import net.minecraftforge.fml.common.Mod.EventBusSubscriber.Bus;
+import electrodynamics.client.misc.SWBFClientExtensions;
+import electrodynamics.common.fluid.SimpleWaterBasedFluidType;
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.api.distmarker.OnlyIn;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.fml.common.EventBusSubscriber;
+import net.neoforged.neoforge.client.event.EntityRenderersEvent;
+import net.neoforged.neoforge.client.extensions.common.RegisterClientExtensionsEvent;
 
 @OnlyIn(Dist.CLIENT)
-@EventBusSubscriber(modid = References.ID, bus = Bus.MOD, value = { Dist.CLIENT })
+@EventBusSubscriber(modid = References.ID, bus = EventBusSubscriber.Bus.MOD, value = { Dist.CLIENT })
 public class ClientRegister {
 
 	public static void setup() {
@@ -23,7 +26,16 @@ public class ClientRegister {
 	@SubscribeEvent
 	@OnlyIn(Dist.CLIENT)
 	public static void registerEntities(EntityRenderersEvent.RegisterRenderers event) {
-		event.registerBlockEntityRenderer(BlastcraftBlockTypes.TILE_CAMOFLAGE.get(), RenderCamoflage::new);
+		event.registerBlockEntityRenderer(BlastcraftTiles.TILE_CAMOFLAGE.get(), RenderCamoflage::new);
+
+	}
+
+	@SubscribeEvent
+	public static void registerClientExtensions(RegisterClientExtensionsEvent event) {
+
+		BlastcraftFluids.FLUIDS.getEntries().forEach((fluid) -> {
+			event.registerFluidType(new SWBFClientExtensions((SimpleWaterBasedFluidType) fluid.get().getFluidType()), fluid.get().getFluidType());
+		});
 
 	}
 
