@@ -6,7 +6,7 @@ import net.minecraft.world.level.block.Blocks;
 public class BlockCustomBricks extends Block {
 
 	public BlockCustomBricks(float hardness, float resistance) {
-		super(Properties.copy(Blocks.IRON_BLOCK).requiresCorrectToolForDrops().strength(hardness, resistance));
+		super(Blocks.IRON_BLOCK.properties().requiresCorrectToolForDrops().strength(hardness, resistance));
 	}
 
 }
