@@ -6,7 +6,7 @@ import net.minecraft.world.level.block.SlabBlock;
 public class BlockCustomSlab extends SlabBlock {
 
 	public BlockCustomSlab(float resistance, float hardness) {
-		super(Properties.copy(Blocks.IRON_BLOCK).requiresCorrectToolForDrops().strength(hardness, resistance));
+		super(Blocks.IRON_BLOCK.properties().requiresCorrectToolForDrops().strength(hardness, resistance));
 	}
 
 }
