@@ -8,8 +8,8 @@ import net.minecraft.world.level.block.state.BlockState;
 
 public class BlockCustomStairs extends StairBlock {
 
-	public BlockCustomStairs(Supplier<BlockState> state, float hardness, float resistance) {
-		super(state, Properties.copy(Blocks.IRON_BLOCK).requiresCorrectToolForDrops().strength(hardness, resistance));
+	public BlockCustomStairs(BlockState state, float hardness, float resistance) {
+		super(state, Blocks.IRON_BLOCK.properties().requiresCorrectToolForDrops().strength(hardness, resistance));
 	}
 
 }
