@@ -1,6 +1,5 @@
 package blastcraft.common.tag;
 
-import blastcraft.common.fluid.types.FluidConcrete;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.BlockTags;
 import net.minecraft.tags.FluidTags;
@@ -36,7 +35,7 @@ public class BlastcraftTags {
 		}
 
 		private static TagKey<Item> forgeTag(String name) {
-			return ItemTags.create(new ResourceLocation("forge", name));
+			return ItemTags.create(ResourceLocation.fromNamespaceAndPath("c", name));
 		}
 
 	}
@@ -54,21 +53,21 @@ public class BlastcraftTags {
 		}
 
 		private static TagKey<Block> forgeTag(String name) {
-			return BlockTags.create(new ResourceLocation("forge", name));
+			return BlockTags.create(ResourceLocation.fromNamespaceAndPath("c", name));
 		}
 
 	}
 
 	public static class Fluids {
 
-		public static final TagKey<Fluid> CONCRETE = forgeTag(FluidConcrete.FORGE_TAG);
+		public static final TagKey<Fluid> CONCRETE = forgeTag("concrete");
 
 		private static void init() {
 
 		}
 
 		private static TagKey<Fluid> forgeTag(String name) {
-			return FluidTags.create(new ResourceLocation("forge", name));
+			return FluidTags.create(ResourceLocation.fromNamespaceAndPath("c", name));
 		}
 
 	}
