@@ -1,27 +1,31 @@
 package blastcraft.common.tile;
 
-import blastcraft.registers.BlastcraftBlockTypes;
+import blastcraft.registers.BlastcraftTiles;
 import electrodynamics.prefab.properties.Property;
-import electrodynamics.prefab.properties.PropertyType;
+import electrodynamics.prefab.properties.PropertyTypes;
 import electrodynamics.prefab.tile.GenericTile;
-import electrodynamics.prefab.tile.components.IComponentType;
 import electrodynamics.prefab.tile.components.type.ComponentPacketHandler;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
+import net.minecraft.world.ItemInteractionResult;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.BlockHitResult;
 
 public class TileCamoflauge extends GenericTile {
 
-	public final Property<BlockState> camoflaugedBlock = property(new Property<>(PropertyType.Blockstate, "camoblock", Blocks.AIR.defaultBlockState()).onChange((prop, oldState) -> {
+	public final Property<BlockState> camoflaugedBlock = property(new Property<>(PropertyTypes.BLOCK_STATE, "camoblock", Blocks.AIR.defaultBlockState()).onChange((prop, oldState) -> {
+		if(level == null) {
+			return;
+		}
 		level.getChunkSource().getLightEngine().checkBlock(worldPosition);
-	}));
+	}).setShouldUpdateOnChange());
 
 	public TileCamoflauge(BlockPos worldPosition, BlockState blockState) {
-		super(BlastcraftBlockTypes.TILE_CAMOFLAGE.get(), worldPosition, blockState);
+		super(BlastcraftTiles.TILE_CAMOFLAGE.get(), worldPosition, blockState);
 		addComponent(new ComponentPacketHandler(this));
 	}
 
@@ -39,16 +43,12 @@ public class TileCamoflauge extends GenericTile {
 	}
 
 	@Override
-	public void onPlace(BlockState oldState, boolean isMoving) {
-		super.onPlace(oldState, isMoving);
-		if (!level.isClientSide) {
-			this.<ComponentPacketHandler>getComponent(IComponentType.PacketHandler).sendProperties();
-		}
-
+	public InteractionResult useWithoutItem(Player player, BlockHitResult hit) {
+		return InteractionResult.PASS;
 	}
 
 	@Override
-	public InteractionResult use(Player arg0, InteractionHand arg1, BlockHitResult arg2) {
-		return InteractionResult.PASS;
+	public ItemInteractionResult useWithItem(ItemStack used, Player player, InteractionHand hand, BlockHitResult hit) {
+		return ItemInteractionResult.PASS_TO_DEFAULT_BLOCK_INTERACTION;
 	}
 }
