@@ -14,7 +14,7 @@ import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.block.SlabBlock;
 import net.minecraft.world.level.block.StairBlock;
 import net.minecraft.world.level.block.WallBlock;
-import net.minecraftforge.common.data.ExistingFileHelper;
+import net.neoforged.neoforge.common.data.ExistingFileHelper;
 
 public class BlastcraftBlockStateProvider extends ElectrodynamicsBlockStateProvider {
 
@@ -30,16 +30,16 @@ public class BlastcraftBlockStateProvider extends ElectrodynamicsBlockStateProvi
 			switch (wall) {
 
 			case base_wall, big_wall, polished_wall, smooth_wall:
-				wallBlock((WallBlock) BlastcraftBlocks.getBlock(wall), blockLoc("blastproofwall/" + name(BlastcraftBlocks.getBlock(wall)).split("_")[0]), true);
+				wallBlock((WallBlock) BlastcraftBlocks.BLOCKS_BLASTPROOFWALL.getValue(wall), blockLoc("blastproofwall/" + name(BlastcraftBlocks.BLOCKS_BLASTPROOFWALL.getValue(wall)).split("_")[0]), true);
 				break;
 			case base_slab, big_slab, polished_slab, smooth_slab:
-				slabBlock((SlabBlock) BlastcraftBlocks.getBlock(wall), blockLoc(name(BlastcraftBlocks.getBlock(wall)).split("_")[0]), blockLoc("blastproofwall/" + name(BlastcraftBlocks.getBlock(wall)).split("_")[0]), true);
+				slabBlock((SlabBlock) BlastcraftBlocks.BLOCKS_BLASTPROOFWALL.getValue(wall), blockLoc(name(BlastcraftBlocks.BLOCKS_BLASTPROOFWALL.getValue(wall)).split("_")[0]), blockLoc("blastproofwall/" + name(BlastcraftBlocks.BLOCKS_BLASTPROOFWALL.getValue(wall)).split("_")[0]), true);
 				break;
 			case base_stairs, big_stairs, polished_stairs, smooth_stairs:
-				stairsBlock((StairBlock) BlastcraftBlocks.getBlock(wall), blockLoc("blastproofwall/" + name(BlastcraftBlocks.getBlock(wall)).split("_")[0]), true);
+				stairsBlock((StairBlock) BlastcraftBlocks.BLOCKS_BLASTPROOFWALL.getValue(wall), blockLoc("blastproofwall/" + name(BlastcraftBlocks.BLOCKS_BLASTPROOFWALL.getValue(wall)).split("_")[0]), true);
 				break;
 			default:
-				simpleBlock(BlastcraftBlocks.getBlock(wall), blockLoc("blastproofwall/" + name(BlastcraftBlocks.getBlock(wall))), true);
+				simpleBlock(BlastcraftBlocks.BLOCKS_BLASTPROOFWALL.getValue(wall), blockLoc("blastproofwall/" + name(BlastcraftBlocks.BLOCKS_BLASTPROOFWALL.getValue(wall))), true);
 			}
 
 		}
@@ -49,16 +49,16 @@ public class BlastcraftBlockStateProvider extends ElectrodynamicsBlockStateProvi
 			switch (wall) {
 
 			case base_wall, big_wall, polished_wall, smooth_wall:
-				wallBlock((WallBlock) BlastcraftBlocks.getBlock(wall), blockLoc("rawblastproofwall/" + name(BlastcraftBlocks.getBlock(wall)).split("_")[0]), true);
+				wallBlock((WallBlock) BlastcraftBlocks.BLOCKS_RAW_BLASTPROOFWALL.getValue(wall), blockLoc("rawblastproofwall/" + name(BlastcraftBlocks.BLOCKS_RAW_BLASTPROOFWALL.getValue(wall)).split("_")[0]), true);
 				break;
 			case base_slab, big_slab, polished_slab, smooth_slab:
-				slabBlock((SlabBlock) BlastcraftBlocks.getBlock(wall), blockLoc(name(BlastcraftBlocks.getBlock(wall)).split("_")[0]), blockLoc("rawblastproofwall/" + name(BlastcraftBlocks.getBlock(wall)).split("_")[0]), true);
+				slabBlock((SlabBlock) BlastcraftBlocks.BLOCKS_RAW_BLASTPROOFWALL.getValue(wall), blockLoc(name(BlastcraftBlocks.BLOCKS_RAW_BLASTPROOFWALL.getValue(wall)).split("_")[0]), blockLoc("rawblastproofwall/" + name(BlastcraftBlocks.BLOCKS_RAW_BLASTPROOFWALL.getValue(wall)).split("_")[0]), true);
 				break;
 			case base_stairs, big_stairs, polished_stairs, smooth_stairs:
-				stairsBlock((StairBlock) BlastcraftBlocks.getBlock(wall), blockLoc("rawblastproofwall/" + name(BlastcraftBlocks.getBlock(wall)).split("_")[0]), true);
+				stairsBlock((StairBlock) BlastcraftBlocks.BLOCKS_RAW_BLASTPROOFWALL.getValue(wall), blockLoc("rawblastproofwall/" + name(BlastcraftBlocks.BLOCKS_RAW_BLASTPROOFWALL.getValue(wall)).split("_")[0]), true);
 				break;
 			default:
-				simpleBlock(BlastcraftBlocks.getBlock(wall), blockLoc("rawblastproofwall/" + name(BlastcraftBlocks.getBlock(wall))), true);
+				simpleBlock(BlastcraftBlocks.BLOCKS_RAW_BLASTPROOFWALL.getValue(wall), blockLoc("rawblastproofwall/" + name(BlastcraftBlocks.BLOCKS_RAW_BLASTPROOFWALL.getValue(wall))), true);
 			}
 		}
 
@@ -67,16 +67,16 @@ public class BlastcraftBlockStateProvider extends ElectrodynamicsBlockStateProvi
 			switch (wall) {
 
 			case base_wall, big_wall, polished_wall, smooth_wall:
-				wallBlock((WallBlock) BlastcraftBlocks.getBlock(wall), blockLoc("carbonplatedwall/" + name(BlastcraftBlocks.getBlock(wall)).split("_")[0]), true);
+				wallBlock((WallBlock) BlastcraftBlocks.BLOCKS_CARBONPLATEDWALL.getValue(wall), blockLoc("carbonplatedwall/" + name(BlastcraftBlocks.BLOCKS_CARBONPLATEDWALL.getValue(wall)).split("_")[0]), true);
 				break;
 			case base_slab, big_slab, polished_slab, smooth_slab:
-				slabBlock((SlabBlock) BlastcraftBlocks.getBlock(wall), blockLoc(name(BlastcraftBlocks.getBlock(wall)).split("_")[0]), blockLoc("carbonplatedwall/" + name(BlastcraftBlocks.getBlock(wall)).split("_")[0]), true);
+				slabBlock((SlabBlock) BlastcraftBlocks.BLOCKS_CARBONPLATEDWALL.getValue(wall), blockLoc(name(BlastcraftBlocks.BLOCKS_CARBONPLATEDWALL.getValue(wall)).split("_")[0]), blockLoc("carbonplatedwall/" + name(BlastcraftBlocks.BLOCKS_CARBONPLATEDWALL.getValue(wall)).split("_")[0]), true);
 				break;
 			case base_stairs, big_stairs, polished_stairs, smooth_stairs:
-				stairsBlock((StairBlock) BlastcraftBlocks.getBlock(wall), blockLoc("carbonplatedwall/" + name(BlastcraftBlocks.getBlock(wall)).split("_")[0]), true);
+				stairsBlock((StairBlock) BlastcraftBlocks.BLOCKS_CARBONPLATEDWALL.getValue(wall), blockLoc("carbonplatedwall/" + name(BlastcraftBlocks.BLOCKS_CARBONPLATEDWALL.getValue(wall)).split("_")[0]), true);
 				break;
 			default:
-				simpleBlock(BlastcraftBlocks.getBlock(wall), blockLoc("carbonplatedwall/" + name(BlastcraftBlocks.getBlock(wall))), true);
+				simpleBlock(BlastcraftBlocks.BLOCKS_CARBONPLATEDWALL.getValue(wall), blockLoc("carbonplatedwall/" + name(BlastcraftBlocks.BLOCKS_CARBONPLATEDWALL.getValue(wall))), true);
 			}
 
 		}
@@ -86,23 +86,23 @@ public class BlastcraftBlockStateProvider extends ElectrodynamicsBlockStateProvi
 			switch (wall) {
 
 			case base_wall, big_wall, polished_wall, smooth_wall:
-				wallBlock((WallBlock) BlastcraftBlocks.getBlock(wall), blockLoc("hardenedbricks/" + name(BlastcraftBlocks.getBlock(wall)).split("_")[0]), true);
+				wallBlock((WallBlock) BlastcraftBlocks.BLOCKS_HARDENEDBRICKS.getValue(wall), blockLoc("hardenedbricks/" + name(BlastcraftBlocks.BLOCKS_HARDENEDBRICKS.getValue(wall)).split("_")[0]), true);
 				break;
 			case base_slab, big_slab, polished_slab, smooth_slab:
-				slabBlock((SlabBlock) BlastcraftBlocks.getBlock(wall), blockLoc(name(BlastcraftBlocks.getBlock(wall)).split("_")[0]), blockLoc("hardenedbricks/" + name(BlastcraftBlocks.getBlock(wall)).split("_")[0]), true);
+				slabBlock((SlabBlock) BlastcraftBlocks.BLOCKS_HARDENEDBRICKS.getValue(wall), blockLoc(name(BlastcraftBlocks.BLOCKS_HARDENEDBRICKS.getValue(wall)).split("_")[0]), blockLoc("hardenedbricks/" + name(BlastcraftBlocks.BLOCKS_HARDENEDBRICKS.getValue(wall)).split("_")[0]), true);
 				break;
 			case base_stairs, big_stairs, polished_stairs, smooth_stairs:
-				stairsBlock((StairBlock) BlastcraftBlocks.getBlock(wall), blockLoc("hardenedbricks/" + name(BlastcraftBlocks.getBlock(wall)).split("_")[0]), true);
+				stairsBlock((StairBlock) BlastcraftBlocks.BLOCKS_HARDENEDBRICKS.getValue(wall), blockLoc("hardenedbricks/" + name(BlastcraftBlocks.BLOCKS_HARDENEDBRICKS.getValue(wall)).split("_")[0]), true);
 				break;
 			default:
-				simpleBlock(BlastcraftBlocks.getBlock(wall), blockLoc("hardenedbricks/" + name(BlastcraftBlocks.getBlock(wall))), true);
+				simpleBlock(BlastcraftBlocks.BLOCKS_HARDENEDBRICKS.getValue(wall), blockLoc("hardenedbricks/" + name(BlastcraftBlocks.BLOCKS_HARDENEDBRICKS.getValue(wall))), true);
 			}
 
 		}
 
 		for (SubtypeWallingGlass glass : SubtypeWallingGlass.values()) {
 
-			glassBlock(BlastcraftBlocks.getBlock(glass), blockLoc("glass/" + glass.tag()), true);
+			glassBlock(BlastcraftBlocks.BLOCKS_WALLINGGLASS.getValue(glass), blockLoc("glass/" + glass.tag()), true);
 
 		}
 
@@ -111,29 +111,29 @@ public class BlastcraftBlockStateProvider extends ElectrodynamicsBlockStateProvi
 			switch (concrete) {
 
 			case bricks_wall, regular_wall, tile_wall:
-				wallBlock((WallBlock) BlastcraftBlocks.getBlock(concrete), blockLoc("concrete/" + name(BlastcraftBlocks.getBlock(concrete)).split("_")[0]), true);
+				wallBlock((WallBlock) BlastcraftBlocks.BLOCKS_CONCRETE.getValue(concrete), blockLoc("concrete/" + name(BlastcraftBlocks.BLOCKS_CONCRETE.getValue(concrete)).split("_")[0]), true);
 				break;
 			case bricks_slab, regular_slab, tile_slab:
-				slabBlock((SlabBlock) BlastcraftBlocks.getBlock(concrete), blockLoc(name(BlastcraftBlocks.getBlock(concrete)).split("_")[0]), blockLoc("concrete/" + name(BlastcraftBlocks.getBlock(concrete)).split("_")[0]), true);
+				slabBlock((SlabBlock) BlastcraftBlocks.BLOCKS_CONCRETE.getValue(concrete), blockLoc(name(BlastcraftBlocks.BLOCKS_CONCRETE.getValue(concrete)).split("_")[0]), blockLoc("concrete/" + name(BlastcraftBlocks.BLOCKS_CONCRETE.getValue(concrete)).split("_")[0]), true);
 				break;
 			case bricks_stairs, regular_stairs, tile_stairs:
-				stairsBlock((StairBlock) BlastcraftBlocks.getBlock(concrete), blockLoc("concrete/" + name(BlastcraftBlocks.getBlock(concrete)).split("_")[0]), true);
+				stairsBlock((StairBlock) BlastcraftBlocks.BLOCKS_CONCRETE.getValue(concrete), blockLoc("concrete/" + name(BlastcraftBlocks.BLOCKS_CONCRETE.getValue(concrete)).split("_")[0]), true);
 				break;
 			default:
-				simpleBlock(BlastcraftBlocks.getBlock(concrete), blockLoc("concrete/" + name(BlastcraftBlocks.getBlock(concrete))), true);
+				simpleBlock(BlastcraftBlocks.BLOCKS_CONCRETE.getValue(concrete), blockLoc("concrete/" + name(BlastcraftBlocks.BLOCKS_CONCRETE.getValue(concrete))), true);
 			}
 
 		}
 
-		simpleBlockCustomRenderType(BlastcraftBlocks.blockCamoflage, blockLoc(name(BlastcraftBlocks.blockCamoflage)), new ResourceLocation("cutout"), true);
+		simpleBlockCustomRenderType(BlastcraftBlocks.BLOCK_CAMOFLAGE, blockLoc(name(BlastcraftBlocks.BLOCK_CAMOFLAGE.get())), ResourceLocation.parse("cutout"), true);
 
-		pressurePlateBlock(BlastcraftBlocks.blockGlassPressurePlate, blockLoc("trans"), new ResourceLocation("cutout"), true);
+		pressurePlateBlock(BlastcraftBlocks.BLOCK_GLASSPRESSUREPLATE.get(), blockLoc("trans"), ResourceLocation.parse("cutout"), true);
 
-		crossBlock(BlastcraftBlocks.blockSpike, blockLoc("spike/" + name(BlastcraftBlocks.blockSpike)), new ResourceLocation("cutout"), true);
-		crossBlock(BlastcraftBlocks.blockSpikeFire, blockLoc("spike/" + name(BlastcraftBlocks.blockSpikeFire)), new ResourceLocation("cutout"), true);
-		crossBlock(BlastcraftBlocks.blockSpikePoison, blockLoc("spike/" + name(BlastcraftBlocks.blockSpikePoison)), new ResourceLocation("cutout"), true);
+		crossBlock(BlastcraftBlocks.BLOCK_SPIKE, blockLoc("spike/" + name(BlastcraftBlocks.BLOCK_SPIKE.get())), ResourceLocation.parse("cutout"), true);
+		crossBlock(BlastcraftBlocks.BLOCK_FIRESPIKE, blockLoc("spike/" + name(BlastcraftBlocks.BLOCK_FIRESPIKE.get())), ResourceLocation.parse("cutout"), true);
+		crossBlock(BlastcraftBlocks.BLOCK_POISONSPIKE, blockLoc("spike/" + name(BlastcraftBlocks.BLOCK_POISONSPIKE.get())), ResourceLocation.parse("cutout"), true);
 
-		horrRotatedBlock(BlastcraftBlocks.blockBlastCompressor, existingBlock(BlastcraftBlocks.blockBlastCompressor), true);
+		horrRotatedBlock(BlastcraftBlocks.BLOCK_BLASTCOMPRESSOR, existingBlock(BlastcraftBlocks.BLOCK_BLASTCOMPRESSOR), true);
 
 	}
 
