@@ -4,7 +4,7 @@ import blastcraft.References;
 import blastcraft.registers.BlastcraftItems;
 import electrodynamics.datagen.client.ElectrodynamicsItemModelsProvider;
 import net.minecraft.data.PackOutput;
-import net.minecraftforge.common.data.ExistingFileHelper;
+import net.neoforged.neoforge.common.data.ExistingFileHelper;
 
 public class BlastcraftItemModelsProvider extends ElectrodynamicsItemModelsProvider {
 
