@@ -8,41 +8,181 @@ import blastcraft.common.block.subtype.SubtypeCarbonPlatedWall;
 import blastcraft.common.block.subtype.SubtypeConcrete;
 import blastcraft.common.block.subtype.SubtypeHardenedBricks;
 import blastcraft.common.block.subtype.SubtypeRawBlastproofWall;
-import blastcraft.common.block.subtype.SubtypeWallingGlass;
 import blastcraft.common.tag.BlastcraftTags;
 import blastcraft.registers.BlastcraftBlocks;
+import electrodynamics.common.block.BlockCustomGlass;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.core.HolderLookup.Provider;
 import net.minecraft.data.PackOutput;
 import net.minecraft.tags.BlockTags;
-import net.minecraftforge.common.data.BlockTagsProvider;
-import net.minecraftforge.common.data.ExistingFileHelper;
+import net.minecraft.world.level.block.Block;
+import net.neoforged.neoforge.common.data.BlockTagsProvider;
+import net.neoforged.neoforge.common.data.ExistingFileHelper;
 
 public class BlastcraftBlockTagsProvider extends BlockTagsProvider {
 
-	public BlastcraftBlockTagsProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> lookupProvider, ExistingFileHelper existingFileHelper) {
-		super(output, lookupProvider, References.ID, existingFileHelper);
-	}
+    public BlastcraftBlockTagsProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> lookupProvider, ExistingFileHelper existingFileHelper) {
+        super(output, lookupProvider, References.ID, existingFileHelper);
+    }
 
-	@Override
-	protected void addTags(Provider provider) {
+    @Override
+    protected void addTags(Provider provider) {
 
-		tag(BlockTags.MINEABLE_WITH_PICKAXE).add(BlastcraftBlocks.blockGlassPressurePlate, BlastcraftBlocks.blockBlastCompressor, BlastcraftBlocks.blockCamoflage, BlastcraftBlocks.blockSpike, BlastcraftBlocks.blockSpikeFire, BlastcraftBlocks.blockSpikePoison).add(BlastcraftBlocks.getAllBlockForSubtype(SubtypeBlastproofWall.values())).add(BlastcraftBlocks.getAllBlockForSubtype(SubtypeRawBlastproofWall.values())).add(BlastcraftBlocks.getAllBlockForSubtype(SubtypeCarbonPlatedWall.values())).add(BlastcraftBlocks.getAllBlockForSubtype(SubtypeHardenedBricks.values())).add(BlastcraftBlocks.getAllBlockForSubtype(SubtypeWallingGlass.values())).add(BlastcraftBlocks.getAllBlockForSubtype(SubtypeConcrete.values()));
+        tag(BlockTags.MINEABLE_WITH_PICKAXE).add(
+                        //
+                        BlastcraftBlocks.BLOCK_GLASSPRESSUREPLATE.get(),
+                        //
+                        BlastcraftBlocks.BLOCK_BLASTCOMPRESSOR.get(),
+                        //
+                        BlastcraftBlocks.BLOCK_CAMOFLAGE.get(),
+                        //
+                        BlastcraftBlocks.BLOCK_SPIKE.get(),
+                        //
+                        BlastcraftBlocks.BLOCK_FIRESPIKE.get(),
+                        //
+                        BlastcraftBlocks.BLOCK_POISONSPIKE.get()
+                        //
+                ).add(BlastcraftBlocks.BLOCKS_BLASTPROOFWALL.getAllValuesArray(new Block[0]))
+                //
+                .add(BlastcraftBlocks.BLOCKS_RAW_BLASTPROOFWALL.getAllValuesArray(new Block[0]))
+                //
+                .add(BlastcraftBlocks.BLOCKS_CARBONPLATEDWALL.getAllValuesArray(new Block[0]))
+                //
+                .add(BlastcraftBlocks.BLOCKS_HARDENEDBRICKS.getAllValuesArray(new Block[0]))
+                //
+                .add(BlastcraftBlocks.BLOCKS_WALLINGGLASS.getAllValuesArray(new BlockCustomGlass[0]))
+                //
+                .add(BlastcraftBlocks.BLOCKS_CONCRETE.getAllValuesArray(new Block[0]));
 
-		tag(BlockTags.NEEDS_STONE_TOOL).add(BlastcraftBlocks.blockGlassPressurePlate, BlastcraftBlocks.blockBlastCompressor, BlastcraftBlocks.blockCamoflage, BlastcraftBlocks.blockSpike, BlastcraftBlocks.blockSpikeFire, BlastcraftBlocks.blockSpikePoison).add(BlastcraftBlocks.getAllBlockForSubtype(SubtypeBlastproofWall.values())).add(BlastcraftBlocks.getAllBlockForSubtype(SubtypeRawBlastproofWall.values())).add(BlastcraftBlocks.getAllBlockForSubtype(SubtypeCarbonPlatedWall.values())).add(BlastcraftBlocks.getAllBlockForSubtype(SubtypeHardenedBricks.values())).add(BlastcraftBlocks.getAllBlockForSubtype(SubtypeWallingGlass.values())).add(BlastcraftBlocks.getAllBlockForSubtype(SubtypeConcrete.values()));
 
-		tag(BlastcraftTags.Blocks.SOLID_BLASTPROOF_WALLS).add(BlastcraftBlocks.getBlock(SubtypeBlastproofWall.base), BlastcraftBlocks.getBlock(SubtypeBlastproofWall.big), BlastcraftBlocks.getBlock(SubtypeBlastproofWall.polished), BlastcraftBlocks.getBlock(SubtypeBlastproofWall.smooth));
+        tag(BlockTags.NEEDS_STONE_TOOL).add(
+                        //
+                        BlastcraftBlocks.BLOCK_GLASSPRESSUREPLATE.get(),
+                        //
+                        BlastcraftBlocks.BLOCK_BLASTCOMPRESSOR.get(),
+                        //
+                        BlastcraftBlocks.BLOCK_CAMOFLAGE.get(),
+                        //
+                        BlastcraftBlocks.BLOCK_SPIKE.get(),
+                        //
+                        BlastcraftBlocks.BLOCK_FIRESPIKE.get(),
+                        //
+                        BlastcraftBlocks.BLOCK_POISONSPIKE.get()
+                        //
+                ).add(BlastcraftBlocks.BLOCKS_BLASTPROOFWALL.getAllValuesArray(new Block[0]))
+                //
+                .add(BlastcraftBlocks.BLOCKS_RAW_BLASTPROOFWALL.getAllValuesArray(new Block[0]))
+                //
+                .add(BlastcraftBlocks.BLOCKS_CARBONPLATEDWALL.getAllValuesArray(new Block[0]))
+                //
+                .add(BlastcraftBlocks.BLOCKS_HARDENEDBRICKS.getAllValuesArray(new Block[0]))
+                //
+                .add(BlastcraftBlocks.BLOCKS_WALLINGGLASS.getAllValuesArray(new BlockCustomGlass[0]))
+                //
+                .add(BlastcraftBlocks.BLOCKS_CONCRETE.getAllValuesArray(new Block[0]));
 
-		tag(BlastcraftTags.Blocks.SOLID_RAW_BLASTPROOF_WALLS).add(BlastcraftBlocks.getBlock(SubtypeRawBlastproofWall.base), BlastcraftBlocks.getBlock(SubtypeRawBlastproofWall.big), BlastcraftBlocks.getBlock(SubtypeRawBlastproofWall.polished), BlastcraftBlocks.getBlock(SubtypeRawBlastproofWall.smooth));
 
-		tag(BlastcraftTags.Blocks.SOLID_CARBON_PLATED_WALLS).add(BlastcraftBlocks.getBlock(SubtypeCarbonPlatedWall.base), BlastcraftBlocks.getBlock(SubtypeCarbonPlatedWall.big), BlastcraftBlocks.getBlock(SubtypeCarbonPlatedWall.polished), BlastcraftBlocks.getBlock(SubtypeCarbonPlatedWall.smooth));
+        tag(BlastcraftTags.Blocks.SOLID_BLASTPROOF_WALLS).add(
+                //
+                BlastcraftBlocks.BLOCKS_BLASTPROOFWALL.getValue(SubtypeBlastproofWall.base),
+                //
+                BlastcraftBlocks.BLOCKS_BLASTPROOFWALL.getValue(SubtypeBlastproofWall.big),
+                //
+                BlastcraftBlocks.BLOCKS_BLASTPROOFWALL.getValue(SubtypeBlastproofWall.polished),
+                //
+                BlastcraftBlocks.BLOCKS_BLASTPROOFWALL.getValue(SubtypeBlastproofWall.smooth)
+                //
+        );
 
-		tag(BlastcraftTags.Blocks.SOLID_HARDENED_BRICKS).add(BlastcraftBlocks.getBlock(SubtypeHardenedBricks.base), BlastcraftBlocks.getBlock(SubtypeHardenedBricks.big), BlastcraftBlocks.getBlock(SubtypeHardenedBricks.polished), BlastcraftBlocks.getBlock(SubtypeHardenedBricks.smooth));
 
-		tag(BlastcraftTags.Blocks.SOLID_CONCRETES).add(BlastcraftBlocks.getBlock(SubtypeConcrete.regular), BlastcraftBlocks.getBlock(SubtypeConcrete.bricks), BlastcraftBlocks.getBlock(SubtypeConcrete.tile));
+        tag(BlastcraftTags.Blocks.SOLID_RAW_BLASTPROOF_WALLS).add(
+                //
+                BlastcraftBlocks.BLOCKS_RAW_BLASTPROOFWALL.getValue(SubtypeRawBlastproofWall.base),
+                //
+                BlastcraftBlocks.BLOCKS_RAW_BLASTPROOFWALL.getValue(SubtypeRawBlastproofWall.big),
+                //
+                BlastcraftBlocks.BLOCKS_RAW_BLASTPROOFWALL.getValue(SubtypeRawBlastproofWall.polished),
+                //
+                BlastcraftBlocks.BLOCKS_RAW_BLASTPROOFWALL.getValue(SubtypeRawBlastproofWall.smooth)
+                //
+        );
 
-		tag(BlockTags.WALLS).add(BlastcraftBlocks.getBlock(SubtypeBlastproofWall.base_wall), BlastcraftBlocks.getBlock(SubtypeBlastproofWall.big_wall), BlastcraftBlocks.getBlock(SubtypeBlastproofWall.polished_wall), BlastcraftBlocks.getBlock(SubtypeBlastproofWall.smooth_wall), BlastcraftBlocks.getBlock(SubtypeRawBlastproofWall.base_wall), BlastcraftBlocks.getBlock(SubtypeRawBlastproofWall.big_wall), BlastcraftBlocks.getBlock(SubtypeRawBlastproofWall.polished_wall), BlastcraftBlocks.getBlock(SubtypeRawBlastproofWall.smooth_wall), BlastcraftBlocks.getBlock(SubtypeCarbonPlatedWall.base_wall), BlastcraftBlocks.getBlock(SubtypeCarbonPlatedWall.big_wall), BlastcraftBlocks.getBlock(SubtypeCarbonPlatedWall.polished_wall), BlastcraftBlocks.getBlock(SubtypeCarbonPlatedWall.smooth_wall), BlastcraftBlocks.getBlock(SubtypeHardenedBricks.base_wall), BlastcraftBlocks.getBlock(SubtypeHardenedBricks.big_wall), BlastcraftBlocks.getBlock(SubtypeHardenedBricks.polished_wall), BlastcraftBlocks.getBlock(SubtypeHardenedBricks.smooth_wall), BlastcraftBlocks.getBlock(SubtypeConcrete.regular_wall), BlastcraftBlocks.getBlock(SubtypeConcrete.bricks_wall), BlastcraftBlocks.getBlock(SubtypeConcrete.tile_wall));
+        tag(BlastcraftTags.Blocks.SOLID_CARBON_PLATED_WALLS).add(
+                //
+                BlastcraftBlocks.BLOCKS_CARBONPLATEDWALL.getValue(SubtypeCarbonPlatedWall.base),
+                //
+                BlastcraftBlocks.BLOCKS_CARBONPLATEDWALL.getValue(SubtypeCarbonPlatedWall.big),
+                //
+                BlastcraftBlocks.BLOCKS_CARBONPLATEDWALL.getValue(SubtypeCarbonPlatedWall.polished),
+                //
+                BlastcraftBlocks.BLOCKS_CARBONPLATEDWALL.getValue(SubtypeCarbonPlatedWall.smooth)
+                //
+        );
 
-	}
+        tag(BlastcraftTags.Blocks.SOLID_HARDENED_BRICKS).add(
+                //
+                BlastcraftBlocks.BLOCKS_HARDENEDBRICKS.getValue(SubtypeHardenedBricks.base),
+                //
+                BlastcraftBlocks.BLOCKS_HARDENEDBRICKS.getValue(SubtypeHardenedBricks.big),
+                //
+                BlastcraftBlocks.BLOCKS_HARDENEDBRICKS.getValue(SubtypeHardenedBricks.polished),
+                //
+                BlastcraftBlocks.BLOCKS_HARDENEDBRICKS.getValue(SubtypeHardenedBricks.smooth)
+                //
+        );
+
+        tag(BlastcraftTags.Blocks.SOLID_CONCRETES).add(
+                //
+                BlastcraftBlocks.BLOCKS_CONCRETE.getValue(SubtypeConcrete.regular),
+                //
+                BlastcraftBlocks.BLOCKS_CONCRETE.getValue(SubtypeConcrete.bricks),
+                //
+                BlastcraftBlocks.BLOCKS_CONCRETE.getValue(SubtypeConcrete.tile)
+                //
+        );
+
+        tag(BlockTags.WALLS).add(
+                //
+                BlastcraftBlocks.BLOCKS_BLASTPROOFWALL.getValue(SubtypeBlastproofWall.base_wall),
+                //
+                BlastcraftBlocks.BLOCKS_BLASTPROOFWALL.getValue(SubtypeBlastproofWall.big_wall),
+                //
+                BlastcraftBlocks.BLOCKS_BLASTPROOFWALL.getValue(SubtypeBlastproofWall.polished_wall),
+                //
+                BlastcraftBlocks.BLOCKS_BLASTPROOFWALL.getValue(SubtypeBlastproofWall.smooth_wall),
+                //
+                BlastcraftBlocks.BLOCKS_RAW_BLASTPROOFWALL.getValue(SubtypeRawBlastproofWall.base_wall),
+                //
+                BlastcraftBlocks.BLOCKS_RAW_BLASTPROOFWALL.getValue(SubtypeRawBlastproofWall.big_wall),
+                //
+                BlastcraftBlocks.BLOCKS_RAW_BLASTPROOFWALL.getValue(SubtypeRawBlastproofWall.polished_wall),
+                //
+                BlastcraftBlocks.BLOCKS_RAW_BLASTPROOFWALL.getValue(SubtypeRawBlastproofWall.smooth_wall),
+                //
+                BlastcraftBlocks.BLOCKS_CARBONPLATEDWALL.getValue(SubtypeCarbonPlatedWall.base_wall),
+                //
+                BlastcraftBlocks.BLOCKS_CARBONPLATEDWALL.getValue(SubtypeCarbonPlatedWall.big_wall),
+                //
+                BlastcraftBlocks.BLOCKS_CARBONPLATEDWALL.getValue(SubtypeCarbonPlatedWall.polished_wall),
+                //
+                BlastcraftBlocks.BLOCKS_CARBONPLATEDWALL.getValue(SubtypeCarbonPlatedWall.smooth_wall),
+                //
+                BlastcraftBlocks.BLOCKS_HARDENEDBRICKS.getValue(SubtypeHardenedBricks.base_wall),
+                //
+                BlastcraftBlocks.BLOCKS_HARDENEDBRICKS.getValue(SubtypeHardenedBricks.big_wall),
+                //
+                BlastcraftBlocks.BLOCKS_HARDENEDBRICKS.getValue(SubtypeHardenedBricks.polished_wall),
+                //
+                BlastcraftBlocks.BLOCKS_HARDENEDBRICKS.getValue(SubtypeHardenedBricks.smooth_wall),
+                //
+                BlastcraftBlocks.BLOCKS_CONCRETE.getValue(SubtypeConcrete.regular_wall),
+                //
+                BlastcraftBlocks.BLOCKS_CONCRETE.getValue(SubtypeConcrete.bricks_wall),
+                //
+                BlastcraftBlocks.BLOCKS_CONCRETE.getValue(SubtypeConcrete.tile_wall)
+                //
+        );
+
+    }
 
 }
