@@ -9,7 +9,7 @@ import net.minecraft.core.HolderLookup;
 import net.minecraft.core.HolderLookup.Provider;
 import net.minecraft.data.PackOutput;
 import net.minecraft.data.tags.FluidTagsProvider;
-import net.minecraftforge.common.data.ExistingFileHelper;
+import net.neoforged.neoforge.common.data.ExistingFileHelper;
 
 public class BlastcraftFluidTagsProvider extends FluidTagsProvider {
 
@@ -20,7 +20,7 @@ public class BlastcraftFluidTagsProvider extends FluidTagsProvider {
 	@Override
 	protected void addTags(Provider provider) {
 
-		tag(BlastcraftTags.Fluids.CONCRETE).add(BlastcraftFluids.fluidConcrete);
+		tag(BlastcraftTags.Fluids.CONCRETE).add(BlastcraftFluids.FLUID_CONCRETE.get());
 
 	}
 
