@@ -1,21 +1,20 @@
 package blastcraft.registers;
 
 import blastcraft.References;
-import blastcraft.common.fluid.types.FluidConcrete;
+import electrodynamics.common.fluid.FluidNonPlaceable;
+import electrodynamics.common.fluid.SimpleWaterBasedFluidType;
+import electrodynamics.prefab.utilities.math.Color;
+import electrodynamics.registers.ElectrodynamicsItems;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.world.level.material.Fluid;
-import net.minecraftforge.registries.DeferredRegister;
-import net.minecraftforge.registries.ForgeRegistries;
+import net.neoforged.neoforge.registries.DeferredHolder;
+import net.neoforged.neoforge.registries.DeferredRegister;
 
 public class BlastcraftFluids {
 
-	public static final DeferredRegister<Fluid> FLUIDS = DeferredRegister.create(ForgeRegistries.FLUIDS, References.ID);
+	public static final DeferredRegister<Fluid> FLUIDS = DeferredRegister.create(Registries.FLUID, References.ID);
 
-	public static FluidConcrete fluidConcrete;
+	public static final DeferredHolder<Fluid, FluidNonPlaceable> FLUID_CONCRETE = FLUIDS.register("fluidconcrete", () -> new FluidNonPlaceable(ElectrodynamicsItems.ITEM_CANISTERREINFORCED, new SimpleWaterBasedFluidType(References.ID, "fluidconcrete", "concrete", Color.WHITE)));
 
-	static {
-
-		FLUIDS.register("fluidconcrete", () -> fluidConcrete = new FluidConcrete());
-
-	}
 
 }
