@@ -1,14 +1,6 @@
 package blastcraft.registers;
 
-import static blastcraft.registers.BlastcraftBlocks.blockBlastCompressor;
-import static blastcraft.registers.BlastcraftBlocks.blockCamoflage;
-import static blastcraft.registers.BlastcraftBlocks.blockGlassPressurePlate;
-import static blastcraft.registers.BlastcraftBlocks.blockSpike;
-import static blastcraft.registers.BlastcraftBlocks.blockSpikeFire;
-import static blastcraft.registers.BlastcraftBlocks.blockSpikePoison;
-
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 
 import blastcraft.References;
@@ -19,99 +11,61 @@ import blastcraft.common.block.subtype.SubtypeHardenedBricks;
 import blastcraft.common.block.subtype.SubtypeRawBlastproofWall;
 import blastcraft.common.block.subtype.SubtypeWallingGlass;
 import blastcraft.prefab.utils.BlastcraftTextUtils;
-import electrodynamics.api.ISubtype;
 import electrodynamics.api.creativetab.CreativeTabSupplier;
+import electrodynamics.api.registration.BulkDeferredHolder;
 import electrodynamics.common.blockitem.types.BlockItemDescriptable;
 import electrodynamics.common.item.ItemDescriptable;
 import net.minecraft.ChatFormatting;
+import net.minecraft.core.registries.Registries;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.event.BuildCreativeModeTabContentsEvent;
-import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.registries.DeferredRegister;
-import net.minecraftforge.registries.ForgeRegistries;
-import net.minecraftforge.registries.RegistryObject;
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.fml.common.EventBusSubscriber;
+import net.neoforged.neoforge.event.BuildCreativeModeTabContentsEvent;
+import net.neoforged.neoforge.registries.DeferredHolder;
+import net.neoforged.neoforge.registries.DeferredRegister;
 
 public class BlastcraftItems {
 
-	public static final DeferredRegister<Item> ITEMS = DeferredRegister.create(ForgeRegistries.ITEMS, References.ID);
+    public static final DeferredRegister<Item> ITEMS = DeferredRegister.create(Registries.ITEM, References.ID);
 
-	public static final HashMap<ISubtype, RegistryObject<Item>> SUBTYPEITEMREGISTER_MAPPINGS = new HashMap<>();
+    public static final BulkDeferredHolder<Item, BlockItemDescriptable, SubtypeHardenedBricks> ITEMS_HARDENEDBRICKS = new BulkDeferredHolder<>(SubtypeHardenedBricks.values(), subtype -> ITEMS.register(subtype.tag(), () -> new BlockItemDescriptable(BlastcraftBlocks.BLOCKS_HARDENEDBRICKS.getValue(subtype), new Item.Properties(), BlastcraftCreativeTabs.MAIN)));
+    public static final BulkDeferredHolder<Item, BlockItemDescriptable, SubtypeConcrete> ITEMS_CONCRETE = new BulkDeferredHolder<>(SubtypeConcrete.values(), subtype -> ITEMS.register(subtype.tag(), () -> new BlockItemDescriptable(BlastcraftBlocks.BLOCKS_CONCRETE.getValue(subtype), new Item.Properties(), BlastcraftCreativeTabs.MAIN)));
+    public static final BulkDeferredHolder<Item, BlockItemDescriptable, SubtypeRawBlastproofWall> ITEMS_RAW_BLASTPROOFWALL = new BulkDeferredHolder<>(SubtypeRawBlastproofWall.values(), subtype -> ITEMS.register(subtype.tag(), () -> new BlockItemDescriptable(BlastcraftBlocks.BLOCKS_RAW_BLASTPROOFWALL.getValue(subtype), new Item.Properties(), BlastcraftCreativeTabs.MAIN)));
+    public static final BulkDeferredHolder<Item, BlockItemDescriptable, SubtypeBlastproofWall> ITEMS_BLASTPROOFWALL = new BulkDeferredHolder<>(SubtypeBlastproofWall.values(), subtype -> ITEMS.register(subtype.tag(), () -> new BlockItemDescriptable(BlastcraftBlocks.BLOCKS_BLASTPROOFWALL.getValue(subtype), new Item.Properties(), BlastcraftCreativeTabs.MAIN)));
+    public static final BulkDeferredHolder<Item, BlockItemDescriptable, SubtypeCarbonPlatedWall> ITEMS_CARBONPLATEDWALL = new BulkDeferredHolder<>(SubtypeCarbonPlatedWall.values(), subtype -> ITEMS.register(subtype.tag(), () -> new BlockItemDescriptable(BlastcraftBlocks.BLOCKS_CARBONPLATEDWALL.getValue(subtype), new Item.Properties(), BlastcraftCreativeTabs.MAIN)));
+    public static final BulkDeferredHolder<Item, BlockItemDescriptable, SubtypeWallingGlass> ITEMS_WALLINGGLASS = new BulkDeferredHolder<>(SubtypeWallingGlass.values(), subtype -> ITEMS.register(subtype.tag(), () -> new BlockItemDescriptable(BlastcraftBlocks.BLOCKS_WALLINGGLASS.getValue(subtype), new Item.Properties(), BlastcraftCreativeTabs.MAIN)));
 
-	static {
-		for (SubtypeBlastproofWall wall : SubtypeBlastproofWall.values()) {
-			SUBTYPEITEMREGISTER_MAPPINGS.put(wall, ITEMS.register(wall.tag(), () -> new BlockItemDescriptable(() -> BlastcraftBlocks.getBlock(wall), new Item.Properties(), () -> BlastcraftCreativeTabs.MAIN.get())));
-		}
+    public static final DeferredHolder<Item, BlockItemDescriptable> ITEM_BLASTCOMPRESSOR = ITEMS.register("blastcompressor", () -> new BlockItemDescriptable(BlastcraftBlocks.BLOCK_BLASTCOMPRESSOR.get(), new Item.Properties(), BlastcraftCreativeTabs.MAIN));
+    public static final DeferredHolder<Item, BlockItemDescriptable> ITEM_CAMOFLAGE = ITEMS.register("camoflage", () -> new BlockItemDescriptable(BlastcraftBlocks.BLOCK_CAMOFLAGE.get(), new Item.Properties(), BlastcraftCreativeTabs.MAIN));
+    public static final DeferredHolder<Item, BlockItemDescriptable> ITEM_GLASSPRESSUREPLATE = ITEMS.register("glasspressureplate", () -> new BlockItemDescriptable(BlastcraftBlocks.BLOCK_GLASSPRESSUREPLATE.get(), new Item.Properties(), BlastcraftCreativeTabs.MAIN));
+    public static final DeferredHolder<Item, BlockItemDescriptable> ITEM_SPIKE = ITEMS.register("spike", () -> new BlockItemDescriptable(BlastcraftBlocks.BLOCK_SPIKE.get(), new Item.Properties(), BlastcraftCreativeTabs.MAIN));
+    public static final DeferredHolder<Item, BlockItemDescriptable> ITEM_FIRESPIKE = ITEMS.register("spikefire", () -> new BlockItemDescriptable(BlastcraftBlocks.BLOCK_FIRESPIKE.get(), new Item.Properties(), BlastcraftCreativeTabs.MAIN));
+    public static final DeferredHolder<Item, BlockItemDescriptable> ITEM_POISONSPIKE = ITEMS.register("spikepoison", () -> new BlockItemDescriptable(BlastcraftBlocks.BLOCK_POISONSPIKE.get(), new Item.Properties(), BlastcraftCreativeTabs.MAIN));
 
-		for (SubtypeRawBlastproofWall wall : SubtypeRawBlastproofWall.values()) {
-			SUBTYPEITEMREGISTER_MAPPINGS.put(wall, ITEMS.register(wall.tag(), () -> new BlockItemDescriptable(() -> BlastcraftBlocks.getBlock(wall), new Item.Properties(), () -> BlastcraftCreativeTabs.MAIN.get())));
-		}
+    public static final DeferredHolder<Item, Item> ITEM_CONCRETEMIX = ITEMS.register("concretemix", () -> new ItemDescriptable(new Item.Properties(), BlastcraftCreativeTabs.MAIN, BlastcraftTextUtils.tooltip("concretemix").withStyle(ChatFormatting.GRAY)));
 
-		for (SubtypeCarbonPlatedWall wall : SubtypeCarbonPlatedWall.values()) {
-			SUBTYPEITEMREGISTER_MAPPINGS.put(wall, ITEMS.register(wall.tag(), () -> new BlockItemDescriptable(() -> BlastcraftBlocks.getBlock(wall), new Item.Properties(), () -> BlastcraftCreativeTabs.MAIN.get())));
-		}
+    @EventBusSubscriber(value = Dist.CLIENT, modid = References.ID, bus = EventBusSubscriber.Bus.MOD)
+    private static class BlastcraftCreativeRegistry {
 
-		for (SubtypeHardenedBricks wall : SubtypeHardenedBricks.values()) {
-			SUBTYPEITEMREGISTER_MAPPINGS.put(wall, ITEMS.register(wall.tag(), () -> new BlockItemDescriptable(() -> BlastcraftBlocks.getBlock(wall), new Item.Properties(), () -> BlastcraftCreativeTabs.MAIN.get())));
-		}
+        @SubscribeEvent
+        public static void registerItems(BuildCreativeModeTabContentsEvent event) {
 
-		for (SubtypeWallingGlass glass : SubtypeWallingGlass.values()) {
+            ITEMS.getEntries().forEach(reg -> {
 
-			SUBTYPEITEMREGISTER_MAPPINGS.put(glass, ITEMS.register(glass.tag(), () -> new BlockItemDescriptable(() -> BlastcraftBlocks.getBlock(glass), new Item.Properties(), () -> BlastcraftCreativeTabs.MAIN.get())));
+                CreativeTabSupplier supplier = (CreativeTabSupplier) reg.get();
 
-		}
+                if (supplier.hasCreativeTab() && supplier.isAllowedInCreativeTab(event.getTab())) {
+                    List<ItemStack> toAdd = new ArrayList<>();
+                    supplier.addCreativeModeItems(event.getTab(), toAdd);
+                    event.acceptAll(toAdd);
+                }
 
-		for (SubtypeConcrete concrete : SubtypeConcrete.values()) {
+            });
 
-			SUBTYPEITEMREGISTER_MAPPINGS.put(concrete, ITEMS.register(concrete.tag(), () -> new BlockItemDescriptable(() -> BlastcraftBlocks.getBlock(concrete), new Item.Properties(), () -> BlastcraftCreativeTabs.MAIN.get())));
+        }
 
-		}
-
-		ITEMS.register("blastcompressor", () -> new BlockItemDescriptable(() -> blockBlastCompressor, new Item.Properties(), () -> BlastcraftCreativeTabs.MAIN.get()));
-		ITEMS.register("camoflage", () -> new BlockItemDescriptable(() -> blockCamoflage, new Item.Properties(), () -> BlastcraftCreativeTabs.MAIN.get()));
-		ITEMS.register("glasspressureplate", () -> new BlockItemDescriptable(() -> blockGlassPressurePlate, new Item.Properties(), () -> BlastcraftCreativeTabs.MAIN.get()));
-		ITEMS.register("spike", () -> new BlockItemDescriptable(() -> blockSpike, new Item.Properties(), () -> BlastcraftCreativeTabs.MAIN.get()));
-		ITEMS.register("spikefire", () -> new BlockItemDescriptable(() -> blockSpikeFire, new Item.Properties(), () -> BlastcraftCreativeTabs.MAIN.get()));
-		ITEMS.register("spikepoison", () -> new BlockItemDescriptable(() -> blockSpikePoison, new Item.Properties(), () -> BlastcraftCreativeTabs.MAIN.get()));
-
-	}
-
-	public static final RegistryObject<Item> ITEM_CONCRETEMIX = ITEMS.register("concretemix", () -> new ItemDescriptable(new Item.Properties(), () -> BlastcraftCreativeTabs.MAIN.get(), BlastcraftTextUtils.tooltip("concretemix").withStyle(ChatFormatting.GRAY)));
-
-	public static Item[] getAllItemForSubtype(ISubtype[] values) {
-		List<Item> list = new ArrayList<>();
-		for (ISubtype value : values) {
-			list.add(SUBTYPEITEMREGISTER_MAPPINGS.get(value).get());
-		}
-		return list.toArray(new Item[] {});
-	}
-
-	public static Item getItem(ISubtype value) {
-		return SUBTYPEITEMREGISTER_MAPPINGS.get(value).get();
-	}
-
-	@Mod.EventBusSubscriber(value = Dist.CLIENT, modid = References.ID, bus = Mod.EventBusSubscriber.Bus.MOD)
-	private static class ElectroCreativeRegistry {
-
-		@SubscribeEvent
-		public static void registerItems(BuildCreativeModeTabContentsEvent event) {
-
-			ITEMS.getEntries().forEach(reg -> {
-
-				CreativeTabSupplier supplier = (CreativeTabSupplier) reg.get();
-
-				if (supplier.hasCreativeTab() && supplier.isAllowedInCreativeTab(event.getTab())) {
-					List<ItemStack> toAdd = new ArrayList<>();
-					supplier.addCreativeModeItems(event.getTab(), toAdd);
-					event.acceptAll(toAdd);
-				}
-
-			});
-
-		}
-
-	}
+    }
 
 }
